+++ CHANGELOG.md
@@ -5,23 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
+## [2.0.0] - 2023-12-09
 
 ### Added
 
 - CHANGELOG.md file
 - 2015-2020 year math problems
 - More descriptions
+- Shuffle
 
 ### Changed
 
 - From Wordpress to React
+- From AWS to GitHub Pages
+- npm to yarn
 
 ### Fixed
 
 - 2021-2023 year math problem sorting
-
-### Removed
 
 ## [1.0.0] - 2023-11-19
 
@@ -30,5 +42,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version with Wordpress
 - HTTPS enabled
 
-[unreleased]: https://github.com/naglissul/vbe-sort/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/naglissul/vbe-sort/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/naglissul/vbe-sort/compare/v2.0.0...v1.0.0
 [1.0.0]: https://github.com/naglissul/vbe-sort/releases/tag/v1.0.0